Fix compiler warnings.

+++ libratbox/src/crypt.c
@@ -1586,7 +1586,7 @@ static char *rb_sha256_crypt_r(const char *key, const char *salt, char *buffer, 
 	rb_sha256_init_ctx(&alt_ctx);
 
 	/* For every character in the password add the entire password.  */
-	for (cnt = 0; cnt < 16 + alt_result[0]; ++cnt)
+	for (cnt = 0; cnt < (size_t)(16 + alt_result[0]); ++cnt)
 		rb_sha256_process_bytes(salt, salt_len, &alt_ctx);
 
 	/* Finish the digest.  */
@@ -2162,7 +2162,7 @@ static char *rb_sha512_crypt_r(const char *key, const char *salt, char *buffer, 
 	rb_sha512_init_ctx(&alt_ctx);
 
 	/* For every character in the password add the entire password.  */
-	for (cnt = 0; cnt < 16 + alt_result[0]; ++cnt)
+	for (cnt = 0; cnt < (size_t)(16 + alt_result[0]); ++cnt)
 		rb_sha512_process_bytes(salt, salt_len, &alt_ctx);
 
 	/* Finish the digest.  */
